added search option

+++ app/Http/Controllers/StaffsController.php
@@ -82,17 +82,19 @@ class StaffsController extends Controller
     }
     public function deleteStaffs(Request $request)
     {
-        $decodedata = json_decode(urldecode($request->input('data')), true);
-        foreach ($decodedata as $id) {
-            $address_id = DB::select('select address_id from staffs where college_id = ?', [$id]);
-            $aid = $address_id[0]->address_id;
-            $address = AddressesModel::find($aid);
-            if (!empty($address)) {
-                $address->delete();
-            }
-            $staff = StaffsModel::find($id);
+        $id = json_decode(urldecode($request->input('data')), true);
+
+        $address_id = DB::select('select address_id from staffs where staff_id = ?', [$id]);
+        $aid = $address_id[0]->address_id;
+        $address = AddressesModel::find($aid);
+        if (!empty($address)) {
+            $address->delete();
+        }
+        $staff = StaffsModel::find($id);
+        if(!empty($staff)) {
             $staff->delete();
         }
+
         return redirect()->back()->with('message', 'deleted staff details suceessfully!');
     }
     public function updateStfForm(Request $request)
