+++ app/Http/Controllers/StudentsController.php
@@ -98,17 +98,19 @@ class StudentsController extends Controller
     }
     public function deleteStudents(Request $request)
     {
-        $decodedata = json_decode(urldecode($request->input('data')), true);
-        foreach ($decodedata as $id) {
-            $address_id = DB::select('select address_id from students where college_id = ?', [$id]);
+        $id = json_decode(urldecode($request->input('data')), true);
+
+            $address_id = DB::select('select address_id from students where student_id = ?', [$id]);
             $aid = $address_id[0]->address_id;
             $address = AddressesModel::find($aid);
             if (!empty($address)) {
                 $address->delete();
             }
             $student = StudentsModel::find($id);
-            $student->delete();
-        }
+            if (!empty($student)) {
+                $student->delete();
+            }
+        
         return redirect()->back()->with('success', 'deleted student details suceessfully!');
     }
     public function updateStdForm(Request $request)
